+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.3.0
 
 - do not include duplicates in imported asset output and total
+- sort imported assets by filesize
+- build button
 
 # 1.2.0